Added cow count bar at top right corner.

+++ src/main/java/edu/chalmers/notenoughspace/event/StorageChangeEvent.java
@@ -12,7 +12,7 @@ public class StorageChangeEvent {
         this.storage = storage;
     }
 
-    public float getNewScore(){
+    public int getNewScore(){
         return storage.calculateScore();
     }
